+++ NEWS.md
@@ -12,6 +12,8 @@ Version 0.15a0 - dev
 - NEW: `Bezier.flattening()` adaptive recursive flattening (approximation)
 - NEW: `Bezier4P.flattening()` adaptive recursive flattening (approximation)
 - NEW: `Path.flattening()` adaptive recursive flattening (approximation)
+- NEW: `Circle.flattening()` approximation determined by a max. sagitta value
+- NEW: `Arc.flattening()` approximation determined by a max. sagitta value
 - NEW: `matplotlib.qsave()`, `ltype` argument to switch between matplotlib dpi 
   based linetype rendering and AutoCAD like drawing units based linetype 
   rendering